remove cloud repo reference
This, it turns out, was a bug.  We got a newer than desired cloud-init
package due to an error in koji tagging.  Now that we've fixed it, the
cloud repo is no longer required and should be removed to avoid
unexpectedly pulling in packages outside of our core atomic/virt tags