ci: Reorganize artifacts into separate dirs: rdgo/ and ostree/
I'd like to add image generation, but it should really be a separate
job in order to avoid blocking everything.  We already should
have ostree composes be async from RPMs.

In order to make this work, we need to cleanly separate the data,
so we can rsync things independently.

I already made this change on the artifacts server by hand, with some
compat symbolic links, so now we need to update the CI jobs.