ci: Add image-cloud, image-pxelive jobs, refactor installer
I was originally thinking we'd just have one "images" job, but after
working on this I think it's better to separate the installer from the
cloud image.  Among other things, if the cloud image fails, we still
want to be able to download/test the ISO by hand.

So refactor things into separate tasks.

I also cleaned up the cloud kickstart to just be `cloud.ks` and fixed
a few bugs in it, and added `pxelive.ks`.

The pxelive job doesn't quite work yet, but let's debug that later.