ci: Dedup code in jobs
Make the "execute script on Duffy node, gathering build logs" a
generic model that's used for everything.  This way we can more
cleanly have logs for failed installer builds for example (not
implemented yet).

Introduce a mechanism to propagate selected environment variables from
Jenkins to the node, which means that the tasks can reference
`$WORKSPACE` as a scratch area.  In the future I plan to also have
things like `${ref}` be parameterized, so that it's easier to
introduce image builds for stable tags.