ci: Rework alpha to be based on recommit promotion (#146)
Previously, we were doing promotion based on updates to the branch
ref.  The problem with this is that we don't have in-repo knowledge of
the history of the target `alpha`. branch.  Things like `ostree log`
show the `continuous` branch.  And in particular, if we want to
generate static deltas between the promoted alpha commits, we'd have
to maintain a branch history of alpha.

So this reworks things to create a new commit, but with the *same
binary content* as the origin commit.  We add metadata `promotion-of`
into the alpha commits that allow us to trace the origin state, and
allows this tool in particular to implement idempotency.

This in turn will allow us to just blindly invoke `ostree static-delta
generate .../alpha` and have it work.

Closes: https://github.com/CentOS/sig-atomic-buildscripts/issues/137