overlay: Nevermind, do manual integration for oci hooks
It turns out docker has a Requires on oci-register-machine with
an Epoch, which ours doesn't include.  This causes libsolv to
depsolve back to an ancient 1.9 docker without it.

We really need to pull docker fully into rdgo to fix issues like
this.  Alternatively, we could teach rdgo to automatically pull
in the upstream Epoch.

In the meantime, this.