centos-ci: add smoketesting for the continuous branch
Add a new job, atomic-tree-smoketest-centos7, which runs after every new
tree compose. This job runs the improved sanity test from the
atomic-host-tests suite:

https://github.com/projectatomic/atomic-host-tests/tree/master/tests/improved-sanity-test

If the test passes, we bump the smoketested rev, which is stored on the
slave's home directory, for lack of a better place. Whenever the
treecompose runs, we pick up the rev and insert it in releases.yaml so
that do-release-tags does the needful.

Disclaimer: About 13.8% of this code is untested.