jjb: transfer repos via rsync --delete
The run-treecompose task edits its `releases.yml` before rsync'ing it to
the build node. However, the `git reset` there would wipe it out. Rather
than doing `git reset`, switch over to `rsync --delete` when
transferring repos so that we're sure the target repo is clean in the
event the node is being reused from a previous task.