centos-ci:  separate 'yum install' operation into two steps
The `atomic*` jobs on ci.centos.org have been failing since Sept 1
because the duffy nodes have not been able to be setup successfully.

The failure was happening when trying to install all the required
packages in one shot, which caused a newer version of `librepo` to be
pulled in from the `atomic7-testing` repo.

But `python-librepo` requires an older version, so we just separate
the 'offending' packages (`libgsystem`, `rpm-ostree`,
`rpm-ostree-toolbox`) into a separate operation.