centos-ci: run the smoketest against continuous branch (#292)
* centos-ci: run the smoketest against continuous branch

We abandoned the `alpha` branch in favor of the `smoketested` branch
(see #270), but the process of running the smoketest has been using
an ancient alpha image for months now.  Building an alpha image
appears to have been broken, so I'm not sure why we continue to try
to test against it.  Additionally, the smoketest has changed
so that we don't need to deploy HEAD-1 before running the test.

* fixup! centos-ci: run the smoketest against continuous branch