centos-ci: handle non 'continuous' branches
The 'atomic-cahc-image-cloud-smoketested' job[1] was failing because it
was expecting a directory named
`/centos-continuous/images-smoketested/cloud` during the `rsync` of
the final artifacts.  This directory was never created as part of the
image building process.

This change should allow the image building process to work with any
non 'continuous' branches going forward.

[1] https://ci.centos.org/job/atomic-cahc-image-cloud-smoketested/1150/console