centos-ci: retain the head commit
In c46bbcf, the `run-tree-smoketest` script was changed to accommodate
the new way of running the a-h-t sanity test.

Unfortunately, this was a bit overzealous, as the HEAD commit is used
later in the script to record which commit was tested.

This change alters the script to pull the commit ID of HEAD, so that
it can be stored again as part of the logs.