jjb: delete no longer defined jobs
These jobs are no longer defined. The new jobs that build the installer
and the cloud image are called e.g. atomic-cahc-installer-continuous.
But anyway, let's just drop it completely for now and let those jobs run
on timer.

(Sidenote: those jobs (and lots of others) still exist and need to be
deleted from Jenkins. Ideally we could use --delete-old, but alas it's
shared with many other tenants. Will do a quick pass after this PR is
merged to clean up what I find.)