libtoolbox.sh: fix cd to nonexistent dir
This is follow-up to #296. We were doing:

    cd images/${imgtype}

which is OK in the continuous case, but in the smoketested case, the
correct dir is `images-smoketested`. This should hopefully make the
smoketested image builder work again.

Also do a few minor tweaks like using `--if-exists` and dropping the
need to specify `--disable-static-deltas`, which should have been long
fixed.