cloud: Bump libvirt memory to 3G
We're hitting issues right now with the installer erroring out. To rule
out whether this is an `ENOSPC` issue, let's try bumping the installer
memory to 3G. If that works, then we can change this override in
rpm-ostree-toolbox (which currently forces it to 2G if it was less), or
make it configurable.