run-tree-smoketest: configure remote temp directory
The 'smoketest' job was failing while running the sanity test because
it was unable to copy a file from the host under test to the host
running the playbook.  The error message suggested to use a temp
directory under `/tmp` as a mitigation.

This creates an `ansible.cfg` in the current directory with the
`remote_tmp` configured.